+++ crates/api-desc/src/crypto.rs
@@ -20,6 +20,8 @@ mod cbc;
 mod ccm;
 #[cfg(feature = "api-crypto-ec")]
 mod ec;
+#[cfg(feature = "api-crypto-ecdsa")]
+mod ecdsa;
 #[cfg(feature = "api-crypto-gcm")]
 mod gcm;
 #[cfg(feature = "internal-api-crypto-hash")]
@@ -37,6 +39,8 @@ pub(crate) fn new() -> Item {
         ccm::new(),
         #[cfg(feature = "api-crypto-ec")]
         ec::new(),
+        #[cfg(feature = "api-crypto-ecdsa")]
+        ecdsa::new(),
         #[cfg(feature = "api-crypto-gcm")]
         gcm::new(),
         #[cfg(feature = "internal-api-crypto-hash")]