+++ crates/runner-host/src/board/crypto.rs
@@ -14,6 +14,8 @@
 
 use wasefire_board_api::crypto::{self, Api};
 
+use crate::board::Board;
+
 pub enum Impl {}
 
 impl Api for Impl {
@@ -23,7 +25,11 @@ impl Api for Impl {
     type HmacSha256 = crypto::SoftwareHmacSha256<Self>;
     type HmacSha384 = crypto::SoftwareHmacSha384<Self>;
     type P256 = crypto::SoftwareP256<Self>;
+    type P256Ecdsa = crypto::SoftwareP256Ecdsa<Self, CryptoRng>;
     type P384 = crypto::SoftwareP384<Self>;
+    type P384Ecdsa = crypto::SoftwareP384Ecdsa<Self, CryptoRng>;
     type Sha256 = crypto::SoftwareSha256;
     type Sha384 = crypto::SoftwareSha384;
 }
+
+type CryptoRng = crypto::CryptoRng<Board>;